+++ commands/audio_effects/chipmunk.js
@@ -27,7 +27,7 @@ module.exports.run = async (client,message,args,prefix,player) => {
             break;
     }
 
-    await queue.setFilters({ effectName: effectModeBoolean });
+    await queue.setFilters({ 'chipmunk': effectModeBoolean });
 
     if (queue.getFiltersEnabled()[0] == effectName || queue.getFiltersEnabled()[0] == undefined) {
         message.reply(`Режим эффекта изменён на | ${effectEnabled}`)