+++ commands/record/record3.js
@@ -36,36 +36,36 @@ const {
 const player = createAudioPlayer();
 
 module.exports.run = async (client,message,args) => {
-    // const channel = message.member?.voice.channel;
-    // if(!channel) return message.channel.send('Join a VC first!');
+    const channel = message.member?.voice.channel;
+    if(!channel) return message.channel.send('Join a VC first!');
 
-    // channel.guild.me.edit({deaf:false})
+    channel.guild.me.edit({deaf:false})
 
-    // const connection = joinVoiceChannel({
-	//     channelId: channel.id,
-	//     guildId: channel.guild.id,
-    //     selfDeaf: false,
-    //     selfMute: false,
-	//     adapterCreator: channel.guild.voiceAdapterCreator,
-    // });
+    const connection = joinVoiceChannel({
+	    channelId: channel.id,
+	    guildId: channel.guild.id,
+        selfDeaf: false,
+        selfMute: false,
+	    adapterCreator: channel.guild.voiceAdapterCreator,
+    });
 
-    // try {
-	// 	await entersState(connection, VoiceConnectionStatus.Ready, 30e3);
-    //     connection.subscribe(player);
-    //     await message.reply('Successfully connected!');
-	// } catch (error) {
-	// 	connection.destroy();
-	// 	throw error;
-	// }
+    try {
+		await entersState(connection, VoiceConnectionStatus.Ready, 30e3);
+        connection.subscribe(player);
+        await message.reply('Successfully connected!');
+	} catch (error) {
+		connection.destroy();
+		throw error;
+	}
 
 
-    // const rec = connection.receiver.subscribe(message.member.id)
+    const rec = connection.receiver.subscribe(message.member.id)
 
-    // const filename = `${Date.now()}-${message.member.id}`;
+    const filename = `${Date.now()}-${message.member.id}`;
 
-    // // const writer = rec.pipe(createWriteStream(join('C:/Users/glebu/Project/red-lamp/recordings', `${filename}.pcm`)))
+    const writer = rec.pipe(createWriteStream(join('C:/Users/glebu/Project/red-lamp/recordings', `${filename}.pcm`)))
 
-    // // let newRecName = `C:/Users/glebu/Project/red-lamp/recordings/end/${Date.now()}newRecording.mp3`
+    // let newRecName = `C:/Users/glebu/Project/red-lamp/recordings/end/${Date.now()}newRecording.mp3`
 
     // function delay(time) {
     //     return new Promise(resolve => setTimeout(resolve, time));
@@ -73,10 +73,15 @@ module.exports.run = async (client,message,args) => {
 
     // delay(10000).then(() => ffmpegconvert(`C:/Users/glebu/Project/red-lamp/recordings/${filename}.pcm`, newRecName))
 
-    // // writer.on('finish', () => {
-    // //     channel.leave();
-    // //     message.channel.send('It went quiet, so I left...');
-    // // });
+    // writer.on('finish', () => {
+    //     channel.leave();
+    //     message.channel.send('It went quiet, so I left...');
+    // });
+
+    // var toWav = require('audiobuffer-to-wav')
+    // var wav = toWav(rec)
+
+    // const wavwriter = wav.pipe(createWriteStream(join('C:/Users/glebu/Project/red-lamp/recordings', `${filename}.wav`)))
 };
 
 module.exports.help = {